# Fix slow performance from recursive `__repr__` and `__hash__` for Go